student CRUD to DTO-based responses with global exception handling

+++ src/main/java/com/example/student_manager/service/StudentService.java
@@ -2,6 +2,7 @@ package com.example.student_manager.service;
 
 import com.example.student_manager.DTO.StudentRequestDTO;
 import com.example.student_manager.DTO.StudentResponseDTO;
+import com.example.student_manager.exception.StudentNotFoundException;
 import com.example.student_manager.model.StudentModel;
 import com.example.student_manager.repository.StudentRepo;
 import org.springframework.stereotype.Service;
@@ -49,18 +50,38 @@ public class StudentService {
                 )).toList();
     }
     // update
-    public StudentModel updateStudent(String id,StudentModel student){
-        StudentModel existingStudent = repository.findById(id)
-                .orElseThrow(()->new RuntimeException("No student found"));
-
+//    public StudentModel updateStudent(String id,StudentModel student){
+//        StudentModel existingStudent = repository.findById(id)
+//                .orElseThrow(()->new RuntimeException("No student found"));
+//
+//        existingStudent.setName(student.getName());
+//        existingStudent.setAge(student.getAge());
+//        existingStudent.setEmail(student.getEmail());
+//
+//        return repository.save(existingStudent);
+//    }
+//
+//    public void deleteStudent(String id){
+//        repository.deleteById(id);
+//    }
+    public StudentResponseDTO updateStudent(String id,StudentResponseDTO student){
+        StudentModel existingStudent=repository.findById(id)
+                .orElseThrow(()->new RuntimeException("NO Student found"));
         existingStudent.setName(student.getName());
         existingStudent.setAge(student.getAge());
         existingStudent.setEmail(student.getEmail());
-
-        return repository.save(existingStudent);
+        StudentModel updated=repository.save(existingStudent);
+        return new StudentResponseDTO(
+                updated.getId(),
+                updated.getName(),
+                updated.getAge(),
+                updated.getEmail()
+        );
     }
-
-    public void deleteStudent(String id){
+    public void delete(String id){
+        if(!repository.existsById(id)){
+            throw new StudentNotFoundException("No student found with this id");
+        }
         repository.deleteById(id);
     }
 }
